requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20231117
torch
pandas>=2.0.0
rich>=13.0.0
click>=8.0.0
jinja2>=3.0.0
flask>=3.0.0
TikTokApi>=7.0.0
playwright>=1.40.0
supadata>=1.0.0
requests>=2.28.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-generativeai>=0.8.0
anthropic>=0.40.0
Pillow>=10.0.0
python-dotenv>=1.0.0
